Update tests and switch import syntax in multiple files
The commit updates various test scripts across the Cypress, Selenium, and Playwright folders. Test scripts are updated with clearer assertions and unnecessary comments were removed. It also changes the syntax from CommonJS require to ECMAScript import in several scripts. Furthermore, a new test script is added for Playwright and "type": "module" is included in the package.json for Cypress.